Rename git to git-commands as terminal gets confused
Cmder and windows terminal throw a script error when in src dir because
of git.js. For some reason the git command tries to execute the file.
